+++ src/components/NftCard.vue
@@ -8,38 +8,7 @@
           href="#"
           @click.prevent="view"
         >
-          <video
-            v-if="nft.media_type?.includes('video') && !loadingMedia"
-            :src="videoUrl"
-            :poster="thumbnailUrl"
-            muted
-            class="img-fluid card-img-top"
-            style="object-fit: cover; height: 100%; object-position: center top"
-          ></video>
-          <img
-            v-else-if="nft.media_type?.includes('image') && !loadingMedia"
-            v-lazy="mediaUrl"
-            style="object-fit: cover; height: 100%; object-position: center top"
-            class="img-fluid card-img-top"
-            alt="Card
-          image cap"
-          />
-          <img
-            v-else-if="loadingMedia"
-            :src="'/loading.gif'"
-            style="object-fit: cover; height: 100%; object-position: center top"
-            class="img-fluid card-img-top"
-            alt="Card
-          image cap"
-          />
-          <img
-            v-else
-            :src="'/thumbnail.jpg'"
-            style="object-fit: cover; height: 100%; object-position: center top"
-            class="img-fluid card-img-top"
-            alt="Card
-          image cap"
-          />
+          <load-media :nft="nft"></load-media>
         </a>
       </figure>
     </template>
@@ -82,11 +51,30 @@
         >
           <strong class="h7 font-weight-bold">Collection </strong><br />
           <div
-            class="d-flex flex-column justify-content-between align-items-center py-2"
+            class="
+              d-flex
+              flex-column
+              justify-content-between
+              align-items-center
+              py-2
+            "
           >
             <div
               v-if="nft.collection.family"
-              class="rounded tex-center d-flex flex-column justify-content-between align-items-center border my-2 w-100 py-1 bg-gradient-primary border-primary"
+              class="
+                rounded
+                tex-center
+                d-flex
+                flex-column
+                justify-content-between
+                align-items-center
+                border
+                my-2
+                w-100
+                py-1
+                bg-gradient-primary
+                border-primary
+              "
               style=""
             >
               <strong class="text-uppercase text-primary small font-weight-bold"
@@ -95,7 +83,20 @@
             </div>
             <div
               v-if="nft.collection.name"
-              class="rounded tex-center d-flex flex-column justify-content-between align-items-center border my-2 w-100 py-1 bg-gradient-primary border-primary"
+              class="
+                rounded
+                tex-center
+                d-flex
+                flex-column
+                justify-content-between
+                align-items-center
+                border
+                my-2
+                w-100
+                py-1
+                bg-gradient-primary
+                border-primary
+              "
               style=""
             >
               <strong class="text-uppercase text-primary small font-weight-bold"
@@ -115,12 +116,31 @@
         <div v-if="nft.attributes && nft.attributes.length" class="mt-2">
           <strong class="h7 font-weight-bold">Attributes </strong><br />
           <div
-            class="d-flex flex-column justify-content-between align-items-center py-2"
+            class="
+              d-flex
+              flex-column
+              justify-content-between
+              align-items-center
+              py-2
+            "
           >
             <div
               v-for="(a, index) in nft.attributes"
               :key="index"
-              class="rounded tex-center d-flex flex-column justify-content-between align-items-center border my-2 w-100 py-1 bg-gradient-primary border-primary"
+              class="
+                rounded
+                tex-center
+                d-flex
+                flex-column
+                justify-content-between
+                align-items-center
+                border
+                my-2
+                w-100
+                py-1
+                bg-gradient-primary
+                border-primary
+              "
               style=""
             >
               <strong
@@ -164,6 +184,7 @@
 <script lang="ts">
 import { computed, defineComponent, ref } from "vue";
 import BaseCard from "@/components/BaseCard.vue";
+import LoadMedia from "@/components/LoadMedia.vue";
 import BaseButton from "@/components/BaseButton.vue";
 import ExternalLink from "@/components/ExternalLink.vue";
 import { useRouter } from "vue-router";
@@ -174,13 +195,13 @@ import {
   getNodeTypeFromNetwork,
 } from "../utils/getNetworkTypeFromCode";
 import { getInspectorUrl } from "../utils/getInspectorUrl";
-import { getIpfsMedia } from "../services/XrpService";
 
 export default defineComponent({
   components: {
     BaseCard,
     BaseButton,
     ExternalLink,
+    LoadMedia,
   },
   props: {
     nft: { type: Object, required: true },
@@ -188,9 +209,7 @@ export default defineComponent({
   setup(props) {
     const router = useRouter();
     const store = useStore();
-    const mediaUrl = ref("");
-    const thumbnailUrl = ref("/loading.gif");
-    const loadingMedia = ref(false);
+
     const network = computed(() => store.getters["user/getNetwork"]);
     const nodetype = computed(() => getNodeTypeFromNetwork(network.value));
 
@@ -222,55 +241,8 @@ export default defineComponent({
         ? props.nft.buyoffers.length
         : 0;
     const countOffers = countSellOffer + countBuyOffer;
-    if (props.nft.url) {
-      if (
-        ["XLS-14", "XLS-16"].includes(props.nft.standard) ||
-        (["XLS-20"].includes(props.nft.standard) &&
-          props.nft.url.split("//")[0] == "https:")
-      ) {
-        mediaUrl.value = props.nft.url || "";
-        thumbnailUrl.value = props.nft.thumbnail || "";
-      } else {
-        loadingMedia.value = true;
-        getIpfsMedia(props.nft.url).then((resp: any) => {
-          loadingMedia.value = false;
-          mediaUrl.value = resp.url;
-        });
-
-        if (props.nft.media_type?.includes("video") && props.nft.thumbnail) {
-          console.log(props.nft.thumbnail);
-          getIpfsMedia(props.nft.thumbnail).then((resp: any) => {
-            thumbnailUrl.value = resp.url;
-          });
-        }
-
-        //  mediaUrl.value = "https://w3s.link/ipfs/" + props.nft.url;
-        // mediaUrl.value = "https://peerkat.mypinata.cloud/ipfs/" + props.nft.url;
-      }
-    }
-
-    // props.nft.url &&
-
-    //   ? props.nft.url
-    //   : props.nft.url
-    //   ? "https://dweb.link/ipfs/" + props.nft.url
-    //   : "";
-
-    // .then((resp: any) => {
-    //   mediaUrl.value = resp.url;
-    //   debugger;
-    // });
-    const videoUrl = computed(() =>
-      props.nft.standard == "XLS-20" && props.nft.thumbnail
-        ? mediaUrl
-        : `${mediaUrl.value}#t=0.5`
-    );
 
     return {
-      mediaUrl,
-      videoUrl,
-      loadingMedia,
-      thumbnailUrl,
       async goToOffer() {
         await store.commit("nft/setCurrent", props.nft);
         router.push({
